Remove rewriting of navigation links
The way we were trying to rewrite navigation links was pretty complex and confusing, although it *kinda* worked.

We don't actually need it, as for MAAS documentation at least, all files are at the same level. So for now, let's just not include it.

We can retrieve it from history if we need it later.