Remove "ignore-files" and add "quiet"
As we now ignore all uppercase filenames, I hope we can simply use that pattern to ignore unneeded files, so I'm removing "ignore-files" option for now. We'll add it back if it's needed.

Now we're skipping files so often, I can see the output getting a bit over the top, so I'm adding a `--quite` option.`