Remove --repository option
There are a few reasons for this:

Firstly, the pulling down of the repository is not something that *needs* to be handled by the builder. It is quite simple for the user to pull down the repository themselves.

Secondly, there was some confusion over the meaning of the "source-path", which was making the code messy. If you are pulling from a repository it meant a different thing than if not.