Change media-folder to media-path
The point of this change is to imply that the media folder is absolute and exists irrepsective of the base repository.

This is because there is a requirement to build each different version branch with the *same* set of media files, and so the media directory shouldn't be considered relative to the branch.